Fixed bug with the hole finding inference
The hole finding inference was not working properly. The preprocessing steps prior to inference (Histogram normalization and cropping) were different than the ones used during training.
This caused the model to not be able to detect holes properly and to return a lot of false positives or no holes at all in certain cases.
The preprocessing steps have been fixed to match the ones used during training.
The model is now able to detect holes properly and returns the expected results.

Version bumped 0.9.4-rc.4