fixed bugs testing workflow in different configuration
Tested where Smartscope runs locally and the AI microservice in a docker container, with the scratch directory bind-mounted to the container.
The paths for scratch being different in the two environments was causing issues. Now, both services get the scratch directory from the SCRATCH_DIR environment variable, which can be set differently in each environment. Logic was changed to keep paths consistend.